+++ pkg/base/curves/bls12381/fieldg2.go
@@ -238,7 +238,7 @@ func (*BaseFieldG2) Select(choice bool, x0, x1 curves.BaseFieldElement) curves.B
 		panic("Not a BLS12381 G2 field element")
 	}
 	return &BaseFieldElementG2{
-		V: new(bimpl.Fp2).CMove(x0p2.V, x1p2.V, utils.BoolTo[int](choice)),
+		V: new(bimpl.Fp2).CMove(x0p2.V, x1p2.V, utils.BoolTo[uint64](choice)),
 	}
 }
